move SW shared code in BaseSession
